moved test/unit/config to test/unit/shared/config
